Add an AMD example
As I suspected you can't use the debug version of the WebGL-Inspector
easily. The non-debug verisons work fine so I don't think this is an
issue.

    twgl-amd.html                 // a normal AMD version
    twgl-amd-embedded.html        // a version that uses embed.js
    twgl-amd-embedded-debug.html  // a version that uses embed.js in debug mode

All of them work in non-debug mode. The last one shows how to get embed.js
to work in debug mode with AMD. Unfortunately the debug version to the extension
doens't work period and I can't think of any easy work around.

Like I said I don't think that's an issue but I thought it was best
to at least check the non-debug version work and document how to
get the embed debug version working